requirements.txt
selenium==3.141.0
requests==2.22.0
aiohttp==3.5.4
beautifulsoup4==4.7.1
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
